fix #949

+++ src/main/java/org/codelibs/fess/app/web/admin/elevateword/EditForm.java
@@ -37,6 +37,6 @@ public class EditForm extends CreateForm {
 
     @Required
     @ValidateTypeFailure
-    public Integer versionNo;
+    public Long versionNo;
 
 }
